+++ packages/wallets/okx/src/custom.d.ts
@@ -2,9 +2,3 @@ declare module '*.svg' {
   const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
   export default content;
 }
-
-declare interface Window {
-  okxwallet: {
-    isOkxWallet: boolean;
-  };
-}