+++ packages/connect-wallet-modal/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @reef-knot/connect-wallet-modal
 
+## 1.1.0
+
+### Patch Changes
+
+- Improve ConnectButtonIconStyle: height, radius
+- Move OKX wallet to the second place in the wallet connection buttons list
+- Updated dependencies
+  - @reef-knot/core-react@1.1.0
+
 ## 1.0.8
 
 ### Patch Changes